src: abstract tracing agent for both legacy and perfetto
