+++ node.gyp
@@ -200,6 +200,7 @@
       'src/timers.cc',
       'src/timer_wrap.cc',
       'src/tracing/agent.cc',
+      'src/tracing/agent_legacy.cc',
       'src/tracing/node_trace_buffer.cc',
       'src/tracing/node_trace_writer.cc',
       'src/tracing/trace_event.cc',
@@ -338,6 +339,7 @@
       'src/tcp_wrap.h',
       'src/timers.h',
       'src/tracing/agent.h',
+      'src/tracing/agent_legacy.h',
       'src/tracing/node_trace_buffer.h',
       'src/tracing/node_trace_writer.h',
       'src/tracing/trace_event.h',
