+++ src/node_v8_platform-inl.h
@@ -4,62 +4,14 @@
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #include <memory>
-#include <string_view>
 
 #include "env-inl.h"
 #include "node.h"
-#include "node_metadata.h"
-#include "node_platform.h"
 #include "node_options.h"
-#include "tracing/node_trace_writer.h"
-#include "tracing/trace_event.h"
-#include "tracing/traced_value.h"
-#include "util.h"
+#include "node_platform.h"
+#include "tracing/agent.h"
 
 namespace node {
-
-// Ensures that __metadata trace events are only emitted
-// when tracing is enabled.
-class NodeTraceStateObserver
-    : public v8::TracingController::TraceStateObserver {
- public:
-  inline void OnTraceEnabled() override {
-    std::string title = GetProcessTitle("");
-    if (!title.empty()) {
-      // Only emit the metadata event if the title can be retrieved
-      // successfully. Ignore it otherwise.
-      TRACE_EVENT_METADATA1(
-          "__metadata", "process_name", "name", TRACE_STR_COPY(title.c_str()));
-    }
-    TRACE_EVENT_METADATA1("__metadata",
-                          "version",
-                          "node",
-                          per_process::metadata.versions.node.c_str());
-    TRACE_EVENT_METADATA1(
-        "__metadata", "thread_name", "name", "JavaScriptMainThread");
-
-    tracing::ProcessMeta trace_process;
-    TRACE_EVENT_METADATA1("__metadata",
-                          "node",
-                          "process",
-                          tracing::CastTracedValue(trace_process));
-    // This only runs the first time tracing is enabled
-    controller_->RemoveTraceStateObserver(this);
-  }
-
-  inline void OnTraceDisabled() override {
-    // Do nothing here. This should never be called because the
-    // observer removes itself when OnTraceEnabled() is called.
-    UNREACHABLE();
-  }
-
-  explicit NodeTraceStateObserver(v8::TracingController* controller)
-      : controller_(controller) {}
-  ~NodeTraceStateObserver() override = default;
-
- private:
-  v8::TracingController* controller_;
-};
 
 struct V8Platform {
   bool initialized_ = false;
@@ -68,20 +20,14 @@ struct V8Platform {
   inline void Initialize(int thread_pool_size) {
     CHECK(!initialized_);
     initialized_ = true;
-    tracing_agent_ = std::make_unique<tracing::Agent>();
-    node::tracing::TraceEventHelper::SetAgent(tracing_agent_.get());
-    node::tracing::TracingController* controller =
-        tracing_agent_->GetTracingController();
-    trace_state_observer_ =
-        std::make_unique<NodeTraceStateObserver>(controller);
-    controller->AddTraceStateObserver(trace_state_observer_.get());
-    tracing_file_writer_ = tracing_agent_->DefaultHandle();
+    tracing_agent_ = tracing::Agent::CreateDefault();
     // Only start the tracing agent if we enabled any tracing categories.
     if (!per_process::cli_options->trace_event_categories.empty()) {
       StartTracingAgent();
     }
     // Tracing must be initialized before platform threads are created.
-    platform_ = new NodePlatform(thread_pool_size, controller);
+    platform_ = new NodePlatform(thread_pool_size,
+                                 tracing_agent_->GetTracingController());
     v8::V8::InitializePlatform(platform_);
   }
   // Make sure V8Platform don not call into Libuv threadpool,
@@ -90,7 +36,6 @@ struct V8Platform {
     if (!initialized_)
       return;
     initialized_ = false;
-    node::tracing::TraceEventHelper::SetAgent(nullptr);
     StopTracingAgent();
     platform_->Shutdown();
     delete platform_;
@@ -98,8 +43,6 @@ struct V8Platform {
     // Destroy tracing after the platform (and platform threads) have been
     // stopped.
     tracing_agent_.reset(nullptr);
-    // The observer remove itself in OnTraceEnabled
-    trace_state_observer_.reset(nullptr);
   }
 
   inline void DrainVMTasks(v8::Isolate* isolate) {
@@ -107,41 +50,19 @@ struct V8Platform {
   }
 
   inline void StartTracingAgent() {
-    constexpr auto convert_to_set =
-        [](auto& categories) -> std::set<std::string> {
-      std::set<std::string> out;
-      for (const auto& s : categories) {
-        out.emplace(std::string(s.data(), s.size()));
-      }
-      return out;
-    };
-    // Attach a new NodeTraceWriter only if this function hasn't been called
-    // before.
-    if (tracing_file_writer_.IsDefaultHandle()) {
-      using std::operator""sv;
-      auto categories = std::views::split(
-          per_process::cli_options->trace_event_categories, ","sv);
-
-      tracing_file_writer_ = tracing_agent_->AddClient(
-          convert_to_set(categories),
-          std::unique_ptr<tracing::AsyncTraceWriter>(
-              new tracing::NodeTraceWriter(
-                  per_process::cli_options->trace_event_file_pattern)),
-          tracing::Agent::kUseDefaultCategories);
-    }
+    if (!initialized_) return;
+    tracing_agent_->StartTracing(
+        per_process::cli_options->trace_event_categories);
   }
 
-  inline void StopTracingAgent() { tracing_file_writer_.reset(); }
-
-  inline tracing::AgentWriterHandle* GetTracingAgentWriter() {
-    return &tracing_file_writer_;
+  inline void StopTracingAgent() {
+    if (!initialized_) return;
+    tracing_agent_->StopTracing();
   }
 
   inline NodePlatform* Platform() { return platform_; }
 
-  std::unique_ptr<NodeTraceStateObserver> trace_state_observer_;
-  std::unique_ptr<tracing::Agent> tracing_agent_;
-  tracing::AgentWriterHandle tracing_file_writer_;
+  std::unique_ptr<tracing::Agent, tracing::Agent::Deleter> tracing_agent_;
   NodePlatform* platform_;
 #else   // !NODE_USE_V8_PLATFORM
   inline void Initialize(int thread_pool_size) {}
@@ -156,22 +77,12 @@ struct V8Platform {
   }
   inline void StopTracingAgent() {}
 
-  inline tracing::AgentWriterHandle* GetTracingAgentWriter() { return nullptr; }
-
   inline NodePlatform* Platform() { return nullptr; }
 #endif  // !NODE_USE_V8_PLATFORM
 };
 
 namespace per_process {
 extern struct V8Platform v8_platform;
-}
-
-inline void StartTracingAgent() {
-  return per_process::v8_platform.StartTracingAgent();
-}
-
-inline tracing::AgentWriterHandle* GetTracingAgentWriter() {
-  return per_process::v8_platform.GetTracingAgentWriter();
 }
 
 inline void DisposePlatform() {
