+++ src/tracing/node_trace_writer.h
@@ -1,11 +1,13 @@
 #ifndef SRC_TRACING_NODE_TRACE_WRITER_H_
 #define SRC_TRACING_NODE_TRACE_WRITER_H_
 
+#if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
+
 #include <sstream>
 #include <queue>
 
 #include "libplatform/v8-tracing.h"
-#include "tracing/agent.h"
+#include "tracing/agent_legacy.h"
 #include "uv.h"
 
 namespace node {
@@ -71,5 +73,7 @@ class NodeTraceWriter : public AsyncTraceWriter {
 
 }  // namespace tracing
 }  // namespace node
+
+#endif  // defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #endif  // SRC_TRACING_NODE_TRACE_WRITER_H_
