+++ src/tracing/trace_event.cc
@@ -4,22 +4,7 @@
 namespace node {
 namespace tracing {
 
-Agent* g_agent = nullptr;
 v8::TracingController* g_controller = nullptr;
-
-void TraceEventHelper::SetAgent(Agent* agent) {
-  if (agent) {
-    g_agent = agent;
-    g_controller = agent->GetTracingController();
-  } else {
-    g_agent = nullptr;
-    g_controller = nullptr;
-  }
-}
-
-Agent* TraceEventHelper::GetAgent() {
-  return g_agent;
-}
 
 v8::TracingController* TraceEventHelper::GetTracingController() {
   return g_controller;
